+++ src/systems/DragAndDropSystem.cpp
@@ -33,9 +33,9 @@ void DragAndDropSystem::update(entityx::EntityManager&, entityx::EventManager&, 
 		//Compute the drag and drop data
 		float delta_x{m_line[1].position.x- m_line[0].position.x};
 		float delta_y{m_line[1].position.y- m_line[0].position.y};
-		double power{hypot(delta_x, delta_y)};//Distance between the two points
-		double angle{atan2(delta_x, delta_y)};//Angle of the line with the horizontal axis
-		angle += b2_pi/2.;//Turn the angle of 90 degrees to fit with the gameplay requirements
+		float power{std::hypot(delta_x, delta_y)};//Distance between the two points
+		float angle{std::atan2(delta_x, delta_y)};//Angle of the line with the horizontal axis
+		angle += b2_pi/2.f;//Turn the angle of 90 degrees to fit with the gameplay requirements
 //		if(angle > b2_pi + b2_pi/4.f)//Keep the angle in the range [-pi, pi]
 //			angle = angle - 2*b2_pi;
 		//Send a command to player's entities to bend them bows according to the drag and drop data
@@ -56,7 +56,7 @@ void DragAndDropSystem::setDragAndDropActivation(bool isActive)
 	{
 		float delta_x{m_line[1].position.x- m_line[0].position.x};
 		float delta_y{m_line[1].position.y- m_line[0].position.y};
-		double angle{atan2(delta_x, delta_y) + b2_pi/2.};//Angle of the line with the horizontal axis
+		float angle{std::atan2(delta_x, delta_y) + b2_pi/2.f};//Angle of the line with the horizontal axis
 		//Shoot the arrow
 		Command shootCommand;
 		shootCommand.targetIsSpecific = false;
